REFACTOR: Plugin Code Migration
