+++ assets/src/components/constants.js
@@ -12,7 +12,7 @@ if ( typeof oneMediaSettings !== 'undefined' ) {
 
 // Export settings as constants.
 export const REST_URL = settings?.restUrl || '';
-export const NONCE = settings?.nonce || '';
+export const NONCE = settings?.restNonce || '';
 export const API_KEY = settings?.apiKey || '';
 export const UPLOAD_NONCE = settings?.uploadNonce || '';
 export const SHARING_AJAX_URL = settings?.ajaxUrl || '';
@@ -34,8 +34,7 @@ export const ONEMEDIA_MEDIA_SHARING = typeof oneMediaMediaSharing !== 'undefined
 export const ONEMEDIA_REST_API_NAMESPACE = 'onemedia';
 export const ONEMEDIA_REST_API_VERSION = 'v1';
 export const ONEMEDIA_REST_API_BASE = '/wp-json/' + ONEMEDIA_REST_API_NAMESPACE + '/' + ONEMEDIA_REST_API_VERSION;
-export const ONEMEDIA_PLUGIN_GOVERNING_SITE = 'governing';
-export const ONEMEDIA_PLUGIN_BRAND_SITE = 'brand';
+export const ONEMEDIA_PLUGIN_GOVERNING_SITE = 'governing-site';
+export const ONEMEDIA_PLUGIN_BRAND_SITE = 'brand-site';
 export const ONEMEDIA_PLUGIN_TAXONOMY_TERM = 'onemedia';
-export const INITIAL_FORM_STATE = { siteName: '', siteUrl: '', apiKey: '' };
 export const MEDIA_PER_PAGE = 12;
