+++ assets/src/components/governing-settings/SiteTable.js
@@ -63,8 +63,8 @@ const SiteTable = ( { sites, onEdit, onDelete, setFormData, setShowModal } ) => 
 						) }
 						{ sites?.map( ( site, index ) => (
 							<tr key={ index }>
-								<td>{ site?.siteName }</td>
-								<td>{ site?.siteUrl }</td>
+								<td>{ site?.name }</td>
+								<td>{ site?.url }</td>
 								<td>
 									<code>
 										{ site.apiKey.substring( 0, 10 ) }...
