+++ inc/Modules/Rest/Basic_Options_Controller.php
@@ -19,6 +19,13 @@ use WP_REST_Server;
 class Basic_Options_Controller extends Abstract_REST_Controller {
 
 	/**
+	 * Health check request timeout.
+	 *
+	 * @var number
+	 */
+	public const HEALTH_CHECK_REQUEST_TIMEOUT = 15;
+
+	/**
 	 * {@inheritDoc}
 	 */
 	public function register_routes(): void {
@@ -366,7 +373,7 @@ class Basic_Options_Controller extends Abstract_REST_Controller {
 		}
 
 		// Check if all the sites for this attachment are connected.
-		$health_check_connected_sites = Utils::health_check_attachment_brand_sites( $attachment_id );
+		$health_check_connected_sites = self::health_check_attachment_brand_sites( $attachment_id );
 
 		return rest_ensure_response( $health_check_connected_sites );
 	}
@@ -377,7 +384,7 @@ class Basic_Options_Controller extends Abstract_REST_Controller {
 	 * @return \WP_REST_Response The response containing the multisite type (single, subdomain or subdirectory).
 	 */
 	public function get_multisite_type(): \WP_REST_Response {
-		$multisite_type = Utils::get_multisite_type();
+		$multisite_type = self::fetch_multisite_type();
 		return new \WP_REST_Response(
 			array(
 				'status'         => 500,
@@ -385,5 +392,222 @@ class Basic_Options_Controller extends Abstract_REST_Controller {
 				'success'        => true,
 			)
 		);
+	}
+
+	/**
+	 * Get the multisite type.
+	 *
+	 * @return string The multisite type (subdomain, subdirectory, or single).
+	 */
+	public function fetch_multisite_type(): string {
+		if ( is_multisite() ) {
+			return is_subdomain_install() ? 'subdomain' : 'subdirectory';
+		}
+		return 'single';
+	}
+
+	/**
+	 * Perform health check on brand sites where a given attachment is shared.
+	 *
+	 * @param int|null $attachment_id The attachment ID.
+	 *
+	 * @return array Array with 'success' boolean and 'failed_sites' array.
+	 */
+	public static function health_check_attachment_brand_sites( int|null $attachment_id ): array {
+		if ( ! $attachment_id ) {
+			return array(
+				'success'      => false,
+				'failed_sites' => array(),
+				'message'      => __( 'Invalid attachment ID.', 'onemedia' ),
+			);
+		}
+
+		// Get URLs of all sites where this attachment is shared.
+		$site_urls = self::get_sync_site_urls_postmeta( $attachment_id );
+
+		if ( empty( $site_urls ) ) {
+			return array(
+				'success'      => true,
+				'failed_sites' => array(),
+				'message'      => __( 'No connected brand sites for this attachment.', 'onemedia' ),
+			);
+		}
+
+		$failed_sites = array();
+		$tracked_urls = array();
+
+		foreach ( $site_urls as $site_url ) {
+			$site_url = rtrim( $site_url, '/' );
+
+			// Skip if we've already processed this URL.
+			if ( in_array( $site_url, $tracked_urls, true ) ) {
+				continue;
+			}
+
+			$tracked_urls[] = $site_url;
+			$api_key        = self::get_brand_site_api_key( $site_url );
+
+			// Brand site not connected.
+			if ( empty( $api_key ) ) {
+				$failed_sites[] = array(
+					'site_name' => self::get_sitename_by_url( $site_url ),
+					'url'       => $site_url,
+					'message'   => __( 'API key not found', 'onemedia' ),
+				);
+				continue;
+			}
+
+			// Perform health check request.
+			$response = wp_remote_get(
+				$site_url . '/wp-json/' . Abstract_REST_Controller::NAMESPACE . '/health-check',
+				array(
+					'timeout' => self::HEALTH_CHECK_REQUEST_TIMEOUT, // phpcs:ignore WordPressVIPMinimum.Performance.RemoteRequestTimeout.timeout_timeout
+					'headers' => array(
+						'X-OneMedia-Token' => $api_key,
+						'Cache-Control'    => 'no-cache, no-store, must-revalidate',
+					),
+				)
+			);
+
+			if ( is_wp_error( $response ) ) {
+				$failed_sites[] = array(
+					'site_name' => self::get_sitename_by_url( $site_url ),
+					'url'       => $site_url,
+					'message'   => $response->get_error_message(),
+				);
+				continue;
+			}
+
+			$response_code = wp_remote_retrieve_response_code( $response );
+			if ( 200 !== $response_code ) {
+				$failed_sites[] = array(
+					'site_name' => self::get_sitename_by_url( $site_url ),
+					'url'       => $site_url,
+					'message'   => sprintf(
+					/* translators: %d is the HTTP response code. */
+						__( 'HTTP %d response', 'onemedia' ),
+						$response_code
+					),
+				);
+			}
+		}
+
+		if ( ! empty( $failed_sites ) ) {
+			$failed_sites_list = array();
+			foreach ( $failed_sites as $failed_site ) {
+				$site_name = $failed_site['site_name'];
+				if ( ! in_array( $site_name, $failed_sites_list, true ) ) {
+					$failed_sites_list[] = $site_name;
+				}
+			}
+
+			return array(
+				'success'      => false,
+				'failed_sites' => $failed_sites,
+				'message'      => sprintf(
+				/* translators: %s is the list of unreachable sites. */
+					__( 'Please check your connection for unreachable sites: %s.', 'onemedia' ),
+					implode( ', ', $failed_sites_list )
+				),
+			);
+		}
+
+		return array(
+			'success'      => true,
+			'failed_sites' => $failed_sites,
+			'message'      => __( 'All connected sites are reachable.', 'onemedia' ),
+		);
+	}
+
+	/**
+	 * Get OneMedia sync site URLs postmeta value.
+	 *
+	 * @param int $attachment_id The attachment ID.
+	 *
+	 * @return array The array of sync site URLs.
+	 */
+	private static function get_sync_site_urls_postmeta( int $attachment_id ): array {
+		$sites = self::get_sync_sites_postmeta( $attachment_id );
+		if ( empty( $sites ) ) {
+			return array();
+		}
+
+		$site_urls = array();
+		foreach ( $sites as $site ) {
+			if ( isset( $site['site'] ) ) {
+				$site_urls[] = untrailingslashit( esc_url_raw( $site['site'] ) );
+			}
+		}
+		return $site_urls;
+	}
+
+	/**
+	 * Get OneMedia sync sites postmeta value.
+	 *
+	 * @param int $attachment_id The attachment ID.
+	 *
+	 * @return array The array of sync sites.
+	 */
+	public static function get_sync_sites_postmeta( int $attachment_id ): array {
+		if ( ! Settings::is_governing_site() || ! $attachment_id ) {
+			return array();
+		}
+
+		$sites = get_post_meta( $attachment_id, Media_Sharing_Controller::ONEMEDIA_SYNC_SITES_POSTMETA_KEY, true );
+		if ( ! is_array( $sites ) ) {
+			return array();
+		}
+		return $sites;
+	}
+
+	/**
+	 * Get saved API key for a given connected brand site URL on governing site.
+	 *
+	 * @param string $site_url The brand site URL.
+	 *
+	 * @return string The saved API key if found, empty string otherwise.
+	 */
+	public static function get_brand_site_api_key( string $site_url ): string {
+		if ( ! Settings::is_governing_site() || empty( $site_url ) ) {
+			return '';
+		}
+
+		$brand_sites = Settings::get_shared_sites();
+		foreach ( $brand_sites as $site ) {
+			if ( rtrim( $site['url'], '/' ) === rtrim( $site_url, '/' ) ) {
+				return $site['api_key'];
+			}
+		}
+		return '';
+	}
+
+	/**
+	 * Get site name by URL.
+	 *
+	 * @param string $site_url The site URL.
+	 *
+	 * @return string The site name if found, empty string otherwise.
+	 */
+	public static function get_sitename_by_url( string $site_url ): string {
+		// If governing site return from option.
+		if ( Settings::is_governing_site() ) {
+			$sites = Settings::get_shared_sites();
+			foreach ( $sites as $site ) {
+				if ( hash_equals( rtrim( $site['url'], '/' ), rtrim( $site_url, '/' ) ) ) {
+					return $site['name'];
+				}
+			}
+		} else {
+			// If brand site create from site_url.
+			$parsed_url = wp_parse_url( $site_url );
+			if ( isset( $parsed_url['host'] ) ) {
+				$host_parts = explode( '.', $parsed_url['host'] );
+				$host_name  = $host_parts[0];
+				$host_name  = str_replace( array( '-', '_' ), ' ', $host_name );
+				$host_name  = ucwords( $host_name );
+				return $host_name;
+			}
+		}
+		return '';
 	}
 }
